Move implementation of ast_{is,get}_token to {is,get}_method.c
