+++ src/ast/is_method.c
@@ -356,3 +356,8 @@ bool ast_is_function_definition(AstRef ast) {
   assert(ast);
   return ast->tag == AST_FUNCTION_DEFINITION;
 }
+
+bool ast_is_token(AstRef ast) {
+  assert(ast);
+  return ast->tag == AST_TOKEN;
+}
